Implement lower_bound with pack instruction
pack instruction is supported in avx2 instruction set.
_mm256_packs_epi32 will saturates and packs 32 bits signed integers a
and b to 16 bits signed integer, but the packing order is not intuitive.
Therefore, it's necessary to permute the order of blocks before packing
it.

Execute scripts/driver.py 24 to observe the results with previos two
version of lower_bound. The latency of pack instruction and movemask
instruction may be different depending on your CPU. It's possible that
the latency of lower_bound_op is lower than lower_bound.